Moved Military Outpost from ships to planets data file
This required showing the "Planets" selector also if the campaign
expansion was not activated (i.e., its checkbox is unchecked).

This required that all planets have a "source" field, which was added
with source="campaign" for the 4 campaign expansion planets. The
Military Outpost already had this from it's time as a unit.

This addresses issue #21